Downgrade unrecognised encoder option from error to warning
The configuration files are written in a configuration language
which has no knowledge of the features macros used in on/off tests.
Therefore, if one disables a tool with associated encoder configuration
options, the encoder will exit with an error due to the unrecognised
configuration options.  This patch makes on/off tests easier by
downgrading the severity of an unrecognised encoder configuration option
from a fatal error to a warning.